+++ rocketpy/simulation/flight.py
@@ -1341,12 +1341,8 @@ class Flight:
             )
             * self.rocket._csys
         )
-        # c = -self.rocket.distance_rocket_nozzle
-        c = (
-            -(self.rocket.nozzle_position - self.rocket.center_of_dry_mass_position)
-            * self.rocket._csys
-        )
-        a = b * Mt / M
+        c = self.rocket.nozzle_to_cdm
+        a = self.rocket.com_to_cdm_function.get_value_opt(t)
         rN = self.rocket.motor.nozzle_radius
         # Prepare transformation matrix
         a11 = 1 - 2 * (e2**2 + e3**2)
@@ -1570,70 +1566,26 @@ class Flight:
         w = Vector([omega1, omega2, omega3])  # Angular velocity vector
 
         # Retrieve necessary quantities
-        rho = self.env.density.get_value_opt(z)
+        ## Rocket mass
         total_mass = self.rocket.total_mass.get_value_opt(t)
         total_mass_dot = self.rocket.total_mass_flow_rate.get_value_opt(t)
-        total_mass_ddot = self.rocket.total_mass_flow_rate.differentiate(t)
+        total_mass_ddot = self.rocket.total_mass_flow_rate.differentiate_complex_step(t)
         ## CM position vector and time derivatives relative to CDM in body frame
-        r_CM_z = (
-            -1
-            * (
-                (
-                    self.rocket.center_of_propellant_position
-                    - self.rocket.center_of_dry_mass_position
-                )
-                * self.rocket._csys
-            )
-            * self.rocket.motor.propellant_mass
-            / total_mass
-        )
-        r_CM = Vector([0, 0, r_CM_z.get_value_opt(t)])
-        r_CM_dot = Vector([0, 0, r_CM_z.differentiate(t)])
+        r_CM_z = self.rocket.com_to_cdm_function
+        r_CM_t = r_CM_z.get_value_opt(t)
+        r_CM = Vector([0, 0, r_CM_t])
+        r_CM_dot = Vector([0, 0, r_CM_z.differentiate_complex_step(t)])
         r_CM_ddot = Vector([0, 0, r_CM_z.differentiate(t, order=2)])
+        ## Nozzle position vector
+        r_NOZ = Vector([0, 0, self.rocket.nozzle_to_cdm])
         ## Nozzle gyration tensor
-        r_NOZ = (
-            -(self.rocket.nozzle_position - self.rocket.center_of_dry_mass_position)
-            * self.rocket._csys
-        )
-        S_noz_33 = 0.5 * self.rocket.motor.nozzle_radius**2
-        S_noz_11 = S_noz_22 = 0.5 * S_noz_33 + 0.25 * r_NOZ**2
-        S_noz_12, S_noz_13, S_noz_23 = 0, 0, 0
-        S_nozzle = Matrix(
-            [
-                [S_noz_11, S_noz_12, S_noz_13],
-                [S_noz_12, S_noz_22, S_noz_23],
-                [S_noz_13, S_noz_23, S_noz_33],
-            ]
-        )
+        S_nozzle = self.rocket.nozzle_gyration_tensor
         ## Inertia tensor
-        I_11 = self.rocket.I_11.get_value_opt(t)
-        I_12 = self.rocket.I_12.get_value_opt(t)
-        I_13 = self.rocket.I_13.get_value_opt(t)
-        I_22 = self.rocket.I_22.get_value_opt(t)
-        I_23 = self.rocket.I_23.get_value_opt(t)
-        I_33 = self.rocket.I_33.get_value_opt(t)
-        I = Matrix(
-            [
-                [I_11, I_12, I_13],
-                [I_12, I_22, I_23],
-                [I_13, I_23, I_33],
-            ]
-        )
+        I = self.rocket.get_inertia_tensor_at_time(t)
         ## Inertia tensor time derivative in the body frame
-        I_11_dot = self.rocket.I_11.differentiate(t)
-        I_12_dot = self.rocket.I_12.differentiate(t)
-        I_13_dot = self.rocket.I_13.differentiate(t)
-        I_22_dot = self.rocket.I_22.differentiate(t)
-        I_23_dot = self.rocket.I_23.differentiate(t)
-        I_33_dot = self.rocket.I_33.differentiate(t)
-        I_dot = Matrix(
-            [
-                [I_11_dot, I_12_dot, I_13_dot],
-                [I_12_dot, I_22_dot, I_23_dot],
-                [I_13_dot, I_23_dot, I_33_dot],
-            ]
-        )
-        ## Inertia tensor relative to CM
+        I_dot = self.rocket.get_inertia_tensor_derivative_at_time(t)
+
+        # Calculate the Inertia tensor relative to CM
         H = (r_CM.cross_matrix @ -r_CM.cross_matrix) * total_mass
         I_CM = I - H
 
@@ -1723,8 +1675,8 @@ class Flight:
                     R1 += comp_lift_xb
                     R2 += comp_lift_yb
                     # Add to total moment
-                    M1 -= (comp_cpz + r_CM_z.get_value_opt(t)) * comp_lift_yb
-                    M2 += (comp_cpz + r_CM_z.get_value_opt(t)) * comp_lift_xb
+                    M1 -= (comp_cpz + r_CM_t) * comp_lift_yb
+                    M2 += (comp_cpz + r_CM_t) * comp_lift_xb
             # Calculates Roll Moment
             try:
                 clf_delta, cld_omega, cant_angle_rad = aero_surface.roll_parameters
@@ -1749,13 +1701,12 @@ class Flight:
                 pass
         weightB = Kt @ Vector([0, 0, -total_mass * self.env.gravity.get_value_opt(z)])
         T00 = total_mass * r_CM
-        r_NOZ_vector = Vector([0, 0, r_NOZ])
-        T03 = 2 * total_mass_dot * (r_NOZ_vector - r_CM) - 2 * total_mass * r_CM_dot
+        T03 = 2 * total_mass_dot * (r_NOZ - r_CM) - 2 * total_mass * r_CM_dot
         T04 = (
-            self.rocket.motor.thrust.get_value_opt(t) * Vector([0, 0, 1])
+            Vector([0, 0, self.rocket.motor.thrust.get_value_opt(t)])
             - total_mass * r_CM_ddot
             - 2 * total_mass_dot * r_CM_dot
-            + total_mass_ddot * (r_NOZ_vector - r_CM)
+            + total_mass_ddot * (r_NOZ - r_CM)
         )
         T05 = total_mass_dot * S_nozzle - I_dot